pyproject.toml
[tool.poetry]
name = "agent-composer"
version = "0.1.0"
description = ""
authors = ["BenderScript <6779302+BenderScript@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pydantic = "^2.7.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
